Move to global type, lots of cleanup, move is-network-error to its own file. remove workflow __error.
