+++ src/lib/stores/error.ts
@@ -1,4 +1,3 @@
-import type { NetworkError } from '$lib/utilities/request-from-api';
 import { writable } from 'svelte/store';
 
 export const networkError = writable<NetworkError | null>(null);
