+++ src/lib/utilities/request-from-api.ts
@@ -30,12 +30,6 @@ export const isTemporalAPIError = (obj: unknown): obj is TemporalAPIError =>
   (obj as TemporalAPIError).message !== undefined &&
   typeof (obj as TemporalAPIError).message === 'string';
 
-export interface NetworkError {
-  statusCode: number;
-  statusText: string;
-  response: Response;
-}
-
 /**
  *  A utility method for making requests to the Temporal API.
  *
