+++ src/routes/__error.svelte
@@ -10,22 +10,19 @@
 </script>
 
 <script lang="ts">
-  import { isNetworkError } from '$lib/utilities/handle-error';
-  import { networkError } from '$lib/stores/error-store';
   import Error from '$lib/components/error.svelte';
+  import { isNetworkError } from '$lib/utilities/is-network-error';
 
   export let error;
   export let status;
   let requestFromAPIError;
-  console.log(error);
+
   try {
     requestFromAPIError = JSON.parse(error.message);
     if (isNetworkError(requestFromAPIError)) {
       status = requestFromAPIError.statusCode;
     }
-  } catch (e) {
-    //
-  }
+  } catch (e) {}
 </script>
 
 <Error {error} />
