+++ src/routes/client-error.svelte
@@ -3,5 +3,5 @@
     statusCode: 404,
     statusText: 'zomg unfound',
     response: {},
-  };
+  } as NetworkError;
 </script>
