+++ src/routes/server-error.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-  import type { NetworkError } from '$lib/utilities/request-from-api';
-
   throw {
     statusCode: 500,
     statusText: "Shits' broke",
